+++ src/Console/KeysCommand.php
@@ -33,8 +33,6 @@ class KeysCommand extends Command
      */
     public function handle(RSA $rsa)
     {
-        $keys = $rsa->createKey($this->input ? (int) $this->option('length') : 4096);
-
         [$publicKey, $privateKey] = [
             Passport::keyPath('oauth-public.key'),
             Passport::keyPath('oauth-private.key'),
@@ -43,6 +41,8 @@ class KeysCommand extends Command
         if ((file_exists($publicKey) || file_exists($privateKey)) && ! $this->option('force')) {
             $this->error('Encryption keys already exist. Use the --force option to overwrite them.');
         } else {
+            $keys = $rsa->createKey($this->input ? (int) $this->option('length') : 4096);
+
             file_put_contents($publicKey, Arr::get($keys, 'publickey'));
             file_put_contents($privateKey, Arr::get($keys, 'privatekey'));