+++ src/PassportServiceProvider.php
@@ -78,12 +78,13 @@ class PassportServiceProvider extends ServiceProvider
      */
     public function register()
     {
+        if (! $this->app->configurationIsCached()) {
+            $this->mergeConfigFrom(__DIR__.'/../config/passport.php', 'passport');
+        }
+
         $this->registerAuthorizationServer();
-
         $this->registerResourceServer();
-
         $this->registerGuard();
-
         $this->offerPublishing();
     }