+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/passport/compare/v7.2.2...master)
+## [Unreleased](https://github.com/laravel/passport/compare/v7.3.2...master)
 
 ### Added
 - Add ability to customize the `RefreshToken` ([#966](https://github.com/laravel/passport/pull/966))
@@ -10,6 +10,27 @@
 - Use diactoros 2.0 and psr-http-factory ([aadf603](https://github.com/laravel/passport/commit/aadf603c1f45cfa4bbf954bfc3abc30cdd572683))
 - Replaced helpers with Blade directives ([#939](https://github.com/laravel/passport/pull/939))
 - Use caret for constraints ([d906804](https://github.com/laravel/passport/commit/d906804c2faccca0333801eccfbf6c2fa5afbaee))
+
+
+## [v7.3.2 (2019-07-11)](https://github.com/laravel/passport/compare/v7.3.1...v7.3.2)
+
+### Fixed
+- Merge default Passport configuration ([#1039](https://github.com/laravel/passport/pull/1039))
+
+
+## [v7.3.1 (2019-07-02)](https://github.com/laravel/passport/compare/v7.3.0...v7.3.1)
+
+### Changed
+- Change server property type in `CheckClientCredentialForAnyScope` ([#1034](https://github.com/laravel/passport/pull/1034))
+
+
+## [v7.3.0 (2019-05-28)](https://github.com/laravel/passport/compare/v7.2.2...v7.3.0)
+
+### Added
+- Allow first party clients to skip the authorization prompt ([#1022](https://github.com/laravel/passport/pull/1022))
+
+### Fixed
+- Fix AccessToken docblock ([#996](https://github.com/laravel/passport/pull/996))
 
 
 ## [v7.2.2 (2019-03-13)](https://github.com/laravel/passport/compare/v7.2.1...v7.2.2)